integrate zustand window size
